# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8080",
            "description": "Generated server url"
        }
    ],
    "tags": [
        {
            "name": "\u7528\u6237",
            "description": "\u7528\u6237\u6ce8\u518c\u7b49"
        },
        {
            "name": "\u516c\u53f8",
            "description": "\u516c\u53f8\u67e5\u8be2\u4e0e CRUD"
        }
    ],
    "paths": {
        "/api/companies/{id}": {
            "put": {
                "tags": [
                    "\u516c\u53f8"
                ],
                "summary": "\u66f4\u65b0\u516c\u53f8",
                "description": "\u516c\u53f8 ID \u653e\u5728\u8def\u5f84\u4e0a\uff0cbody \u4e2d\u4e0d\u9700\u8981 id",
                "operationId": "updateCompany",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "\u516c\u53f8 ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 6
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CompanyUpdateDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "\u516c\u53f8"
                ],
                "summary": "\u5220\u9664\u516c\u53f8",
                "operationId": "deleteCompany",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "\u516c\u53f8 ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 6
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultInteger"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/update-avatar": {
            "post": {
                "tags": [
                    "\u7528\u6237"
                ],
                "summary": "\u66f4\u65b0\u7528\u6237\u5934\u50cf",
                "description": "\u9700\u5728 Header \u6dfb\u52a0 Authorization: Bearer \u767b\u5f55\u8fd4\u56de\u7684 token",
                "operationId": "updateAvatar",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "query",
                        "description": "\u7528\u6237 ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "avatarUrl",
                        "in": "query",
                        "description": "\u5934\u50cf URL",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/register": {
            "post": {
                "tags": [
                    "\u7528\u6237"
                ],
                "summary": "\u7528\u6237\u6ce8\u518c",
                "description": "\u7528\u6237\u540d\u4e0d\u53ef\u91cd\u590d\uff1b\u6210\u529f\u8fd4\u56de\u65b0\u7528\u6237 ID",
                "operationId": "register",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRegisterDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultLong"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/login": {
            "post": {
                "tags": [
                    "\u7528\u6237"
                ],
                "operationId": "login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserLoginDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultLoginResponseDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/avatar": {
            "post": {
                "tags": [
                    "\u7528\u6237"
                ],
                "summary": "\u4e0a\u4f20\u5934\u50cf",
                "description": "\u4e0a\u4f20\u5934\u50cf\u6587\u4ef6\uff0c\u6210\u529f\u540e\u8fd4\u56de\u5934\u50cf\u8bbf\u95ee\u8def\u5f84",
                "operationId": "uploadAvatar",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "query",
                        "description": "\u7528\u6237 ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary",
                                        "description": "\u5934\u50cf\u6587\u4ef6"
                                    }
                                },
                                "required": [
                                    "file"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultString"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/companies/addCompany": {
            "post": {
                "tags": [
                    "\u516c\u53f8"
                ],
                "summary": "\u65b0\u589e\u516c\u53f8",
                "operationId": "addCompany",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CompanyAddDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultCompany"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/company": {
            "get": {
                "tags": [
                    "\u516c\u53f8"
                ],
                "summary": "\u793a\u4f8b\u516c\u53f8\uff08\u6f14\u793a\uff09",
                "description": "\u8fd4\u56de\u786c\u7f16\u7801 Map\uff0c\u4ec5\u7528\u4e8e\u5b66\u4e60\u6f14\u793a",
                "operationId": "getCompany",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {}
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/company/{id}": {
            "get": {
                "tags": [
                    "\u516c\u53f8"
                ],
                "summary": "\u6309 ID \u67e5\u8be2\u516c\u53f8",
                "operationId": "getCompanyById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "\u516c\u53f8 ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultCompanyResponseDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/companies": {
            "get": {
                "tags": [
                    "\u516c\u53f8"
                ],
                "summary": "\u83b7\u53d6\u516c\u53f8\u5217\u8868",
                "description": "\u5206\u9875\u67e5\u8be2\uff0c\u652f\u6301\u6309\u57ce\u5e02\u3001\u662f\u5426\u53cc\u4f11\u3001\u540d\u79f0\u5173\u952e\u8bcd\u7b5b\u9009",
                "operationId": "getCompanyList",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "description": "\u67e5\u8be2\u6761\u4ef6\uff08Query \u53c2\u6570\uff09",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/CompanyQueryDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultPageResultCompany"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "CompanyUpdateDTO": {
                "type": "object",
                "description": "\u8981\u66f4\u65b0\u7684\u516c\u53f8\u4fe1\u606f",
                "properties": {
                    "name": {
                        "type": "string",
                        "description": "\u516c\u53f8\u540d\u79f0",
                        "example": "\u5468\u5927\u6f20\u516c\u53f8",
                        "maxLength": 100,
                        "minLength": 0
                    },
                    "city": {
                        "type": "string",
                        "description": "\u6240\u5728\u57ce\u5e02",
                        "example": "\u4e0a\u6d77",
                        "minLength": 1
                    },
                    "weekendsOff": {
                        "type": "boolean",
                        "description": "\u662f\u5426\u53cc\u4f11",
                        "example": true
                    }
                },
                "required": [
                    "city",
                    "name",
                    "weekendsOff"
                ]
            },
            "ResultVoid": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {}
                }
            },
            "UserRegisterDTO": {
                "type": "object",
                "description": "\u6ce8\u518c\u4fe1\u606f",
                "properties": {
                    "username": {
                        "type": "string",
                        "description": "\u7528\u6237\u540d\uff0c3-20 \u4f4d",
                        "example": "zhangdaya",
                        "maxLength": 20,
                        "minLength": 3
                    },
                    "password": {
                        "type": "string",
                        "description": "\u5bc6\u7801\uff0c6-30 \u4f4d",
                        "example": 123456,
                        "maxLength": 30,
                        "minLength": 6
                    },
                    "nickname": {
                        "type": "string",
                        "description": "\u6635\u79f0\uff08\u53ef\u9009\uff09",
                        "example": "\u5927\u7259"
                    }
                },
                "required": [
                    "password",
                    "username"
                ]
            },
            "ResultLong": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "UserLoginDTO": {
                "type": "object",
                "description": "\u767b\u5f55\u4fe1\u606f",
                "properties": {
                    "username": {
                        "type": "string",
                        "description": "\u7528\u6237\u540d\uff0c3-20 \u4f4d",
                        "example": "zhangdaya",
                        "maxLength": 20,
                        "minLength": 3
                    },
                    "password": {
                        "type": "string",
                        "description": "\u5bc6\u7801\uff0c6-30 \u4f4d",
                        "example": 123456,
                        "maxLength": 30,
                        "minLength": 6
                    }
                },
                "required": [
                    "password",
                    "username"
                ]
            },
            "LoginResponseDTO": {
                "type": "object",
                "description": "\u767b\u5f55\u6210\u529f\u54cd\u5e94",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "description": "\u7528\u6237 ID",
                        "example": 1
                    },
                    "username": {
                        "type": "string",
                        "description": "\u7528\u6237\u540d",
                        "example": "zhangdaya"
                    },
                    "avatar": {
                        "type": "string",
                        "description": "\u5934\u50cf URL",
                        "example": "https://example.com/avatar.png"
                    },
                    "token": {
                        "type": "string",
                        "description": "\u767b\u5f55\u4ee4\u724c\uff08\u540e\u7eed\u9274\u6743\u7528\uff09"
                    }
                }
            },
            "ResultLoginResponseDTO": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/LoginResponseDTO"
                    }
                }
            },
            "ResultString": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "string"
                    }
                }
            },
            "CompanyAddDTO": {
                "type": "object",
                "description": "\u65b0\u589e\u516c\u53f8\u4fe1\u606f",
                "properties": {
                    "name": {
                        "type": "string",
                        "description": "\u516c\u53f8\u540d\u79f0",
                        "example": "\u5f20\u5927\u7259\u79d1\u6280\u6709\u9650\u516c\u53f8",
                        "maxLength": 100,
                        "minLength": 0
                    },
                    "city": {
                        "type": "string",
                        "description": "\u6240\u5728\u57ce\u5e02",
                        "example": "\u821f\u5c71",
                        "minLength": 1
                    },
                    "weekendsOff": {
                        "type": "boolean",
                        "description": "\u662f\u5426\u53cc\u4f11",
                        "example": true
                    }
                },
                "required": [
                    "city",
                    "name",
                    "weekendsOff"
                ]
            },
            "Company": {
                "type": "object",
                "description": "\u516c\u53f8\u4fe1\u606f",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "description": "\u516c\u53f8ID",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "description": "\u516c\u53f8\u540d\u79f0",
                        "example": "\u516c\u53f81",
                        "maxLength": 100,
                        "minLength": 2
                    },
                    "weekendsOff": {
                        "type": "boolean",
                        "description": "\u662f\u5426\u53cc\u4f11",
                        "example": true
                    },
                    "city": {
                        "type": "string",
                        "description": "\u57ce\u5e02",
                        "example": "\u5317\u4eac",
                        "minLength": 1
                    }
                },
                "required": [
                    "city",
                    "name",
                    "weekendsOff"
                ]
            },
            "ResultCompany": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/Company"
                    }
                }
            },
            "CompanyResponseDTO": {
                "type": "object",
                "description": "\u516c\u53f8\u4fe1\u606f\uff08\u54cd\u5e94\uff09",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "description": "\u516c\u53f8 ID",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "description": "\u516c\u53f8\u540d\u79f0",
                        "example": "\u5b81\u6ce2\u6d4b\u8bd5\u79d1\u6280\u6709\u9650\u516c\u53f8"
                    },
                    "city": {
                        "type": "string",
                        "description": "\u6240\u5728\u57ce\u5e02",
                        "example": "\u5b81\u6ce2"
                    },
                    "weekendsOff": {
                        "type": "boolean",
                        "description": "\u662f\u5426\u53cc\u4f11",
                        "example": true
                    }
                }
            },
            "ResultCompanyResponseDTO": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/CompanyResponseDTO"
                    }
                }
            },
            "CompanyQueryDTO": {
                "type": "object",
                "description": "\u516c\u53f8\u5217\u8868\u67e5\u8be2\u53c2\u6570\uff08Query\uff09",
                "properties": {
                    "page": {
                        "type": "integer",
                        "format": "int32",
                        "default": 1,
                        "description": "\u9875\u7801\uff0c\u4ece 1 \u5f00\u59cb",
                        "example": 1
                    },
                    "pageSize": {
                        "type": "integer",
                        "format": "int32",
                        "default": 10,
                        "description": "\u6bcf\u9875\u6761\u6570",
                        "example": 10
                    },
                    "city": {
                        "type": "string",
                        "description": "\u57ce\u5e02\uff08\u7cbe\u786e\u5339\u914d\uff09",
                        "example": "\u5b81\u6ce2"
                    },
                    "weekendsOff": {
                        "type": "boolean",
                        "description": "\u662f\u5426\u53cc\u4f11",
                        "example": true
                    },
                    "keyword": {
                        "type": "string",
                        "description": "\u516c\u53f8\u540d\u79f0\u5173\u952e\u8bcd\uff08\u6a21\u7cca\u5339\u914d\uff09",
                        "example": "\u79d1\u6280"
                    }
                }
            },
            "PageResultCompany": {
                "type": "object",
                "properties": {
                    "list": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Company"
                        }
                    },
                    "total": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "page": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "pageSize": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
            "ResultPageResultCompany": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/PageResultCompany"
                    }
                }
            },
            "ResultInteger": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            }
        }
    }
}